+++ frontend/src/selectors/getTargetNetwork.js
@@ -1,0 +1,1 @@
+export const getTargetNetwork = (state) => state.globalProperties.network;