+++ frontend/src/utils/web3old.js
@@ -1,0 +1,3 @@
+export const web3old = window.web3;
+
+export default web3old;