+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OPEN API Keys management
 - Scaffold templates functionality
 - Functionality for updating Scaffold Information
+- Scaffold activation via MetaMask
 
 ## [2.0.0] - 2018-06-04
 ### Added